Add files via upload
This analysis provides valuable insights into the passenger and flight data contained in the "Airline Dataset Updated - v2.csv." It is important to note the diversity in passenger nationalities and the varying flight statuses in the dataset. The report sheds light on the oldest and youngest passengers and provides context regarding the most recent flight departures. The dataset's composition with passengers from various countries adds to its richness.